pyproject.toml
[project]
name = "image-contest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "open-clip-torch>=2.32.0",
    "openai>=1.84.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "streamlit>=1.45.1",
    "torch>=2.7.1",
    "click>=8.0.0,<9.0.0",
]
